Add a GBA cpu pointer to the memory map
This allows us to pass the memory map to all memory read/write
functions and still be able to access the CPU if needed (for
example when determining if a BIOS address read is allowed).
This pattern will also useful when implementing I/O registers
stored in their corresponding components (PPU, APU, etc).