Implement support for loading a GBA BIOS file
This will help me not have to fully emulate the BIOS
system calls yet, though support for this should
be added in the future. If no BIOS file is provided,
fall back to emulating BIOS calls.

Continue to skip the startup BIOS code until enough
of the system is implemented for it to run correctly.